Android instructions fixed correctly
Previewed changes and they look ok now. 

The numbers on the NPM readme page on Chrome were coming out incorrectly, although they are OK on github.  
I checked the history of the file and this was never changed so it is an issue.  

Instead of 1, two indented bullets and then 2 and 3, it has: 1, two main bullets BEFORE the numbers , then again 1 and 2. 

I fixed that by inserting the numbers manually. Note 4 spaces at end of one of the lines, for a line-break. 

I also gave an example of how to add the packages to the `getPackage()` list, noting the needed comma.